requirements.txt
PyQt6
pyqt6-tools
PySide6
pyqtgraph
pyserial
numpy
scipy
matplotlib